feat: udistrital/sga_documentacion#180 Refactorización de la estructura de carpetas de legalización
