--- src/app/modules/legalizacion-matricula/components/modal/modal.component.scss
@@ -1,60 +1,0 @@
-.modal-card-container {
-  width: 800px; /* Definir el ancho deseado */
-  height: 800px; /* Definir la altura deseada */
-  padding: 20px;
-}
-
-.modal-card {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  padding: 20px;
-}
-
-.document-viewer {
-  flex: 1;
-}
-
-.input-section {
-  flex: 1;
-  display: flex;
-  height:70vh;
-  flex-direction: column;
-  gap: 10px;
-}
-
-.boton-aprobar {
-  border: solid 1px #03678f;
-  background-color: #03678f;
-  border-color: #03678f;
-  cursor: pointer;
-  color: white !important;
-  display: flex;
-}
-
-.boton-rechazar {
-  border: solid 1px var(--danger-accent);
-  background-color: var(--danger-accent);
-  border-color: var(--danger-accent);
-  cursor: pointer;
-  color: white !important;
-  display: flex;
-}
-
-.flex-item {
-  flex-grow: 1;
-}
-
-.input-container {
-  display: flex;
-  flex-direction: column;
-}
-
-.input-container label {
-  margin-bottom: 5px;
-}
-
-.button-container {
-  display: flex;
-  gap: 10px;
-}
