+++ src/app/modules/legalizacion/legalizacion-matricula/legalizacion-matricula.component.html
@@ -403,3 +403,18 @@
     </mat-card-content>
   </mat-card>
 </div>
+
+<div *ngIf="!estaAutorizado">
+  <mat-card>
+    <mat-card-content>
+        <mat-card class="contendedor-no-autorizado">
+            <mat-card-header>
+                <h1>{{'ERROR.rol_insuficiente_titulo' | translate}}</h1>
+            </mat-card-header>
+            <mat-card-content>
+                <p>{{'ERROR.rol_insuficiente' | translate}}</p>
+            </mat-card-content>    
+        </mat-card>
+    </mat-card-content>
+  </mat-card>
+</div>
